refactor(create-webstone-app): move app logic to file
